Refactor UI into modular classes
